+++ src/components/Pages/About.js
@@ -9,21 +9,18 @@ import {
   Slider,
   SwitchPosition,
 } from "../Switch/Switch";
+import { P, HeroContainer, H1, H2 } from "../HeroSection/HeroSection";
 import {
-  P,
   ImgAbout,
-  HeroContainer,
-  H1,
-  H2,
-  ImgContainer,
   IoJS,
   Html5,
   React,
   Container,
   Graduation,
   Tecnologies,
-  GraduationCap
-} from "../HeroSection/HeroSection";
+  GraduationCap,
+} from "../HeroSectionAbout/HeroSectionAbout";
+
 import nightGirl from "../../Images/universeGirl.svg";
 import myselfPhoto from "../../Images/myself_pictures.jpeg";
 
@@ -37,7 +34,7 @@ const About = () => {
   const text = `${
     theme === "light"
       ? `${"I love creating new things and getting into new projects, an exceptional team player, willing to maintain extensive communication, and always give my best."}`
-      : "hello my name is Myllena"
+      : "Night theme: I love creating new things and getting into new projects, an exceptional team player, willing to maintain extensive communication, and always give my best."
   }`;
 
   return (
@@ -56,26 +53,29 @@ const About = () => {
             <P>{text}</P>
             <Container>
               <Graduation>
-              <H2>Graduation</H2>
-              <GraduationCap />
-              <P>
-                Information Systems 2018-2023 State University of Montes Claros
-              </P></Graduation><Tecnologies>
-              <H2>Technologies</H2>
-              <P>Some technologies that I like to work:</P>
-              <React />
-              <IoJS />
-              <Html5 /></Tecnologies>
-              <ImgContainer>
-                <ImgAbout
-                  src={theme === "light" ? myselfPhoto : nightGirl}
-                  alt={
-                    theme === "light"
-                      ? "This Káren Myllena Gonçalves"
-                      : "Girl with universe on her hair"
-                  }
-                />
-              </ImgContainer>
+                <H2>Graduation</H2>
+                <GraduationCap />
+                <P>
+                  Information Systems 2018-2023 State University of Montes
+                  Claros
+                </P>
+              </Graduation>
+              <Tecnologies>
+                <H2>Technologies</H2>
+                <P>Some technologies that I like to work:</P>
+                <React />
+                <IoJS />
+                <Html5 />
+              </Tecnologies>
+
+              <ImgAbout
+                src={theme === "light" ? myselfPhoto : nightGirl}
+                alt={
+                  theme === "light"
+                    ? "This Káren Myllena Gonçalves"
+                    : "Girl with universe on her hair"
+                }
+              />
             </Container>
           </HeroContainer>
         </ThemeChanger>
